+++ app/components/primer/primer.ts
@@ -9,7 +9,7 @@ import './alpha/x_banner'
 import './beta/auto_complete/auto_complete'
 import './beta/clipboard_copy'
 import './local_time'
-import './tab_container_component'
+import './alpha/tab_container'
 import './time_ago_component'
 import '../../../lib/primer/forms/primer_multi_input'
 import '../../../lib/primer/forms/primer_text_field'
