+++ lib/tasks/docs.rake
@@ -76,7 +76,7 @@ namespace :docs do
       Primer::Beta::State,
       Primer::Beta::Spinner,
       Primer::SubheadComponent,
-      Primer::TabContainerComponent,
+      Primer::Alpha::TabContainer,
       Primer::Beta::Text,
       Primer::Alpha::TextField,
       Primer::TimeAgoComponent,
@@ -106,7 +106,7 @@ namespace :docs do
       Primer::Beta::AutoComplete,
       Primer::Alpha::Banner,
       Primer::Beta::ClipboardCopy,
-      Primer::TabContainerComponent,
+      Primer::Alpha::TabContainer,
       Primer::TimeAgoComponent,
       Primer::Alpha::UnderlinePanels,
       Primer::Alpha::TabPanels,
