+++ static/arguments.json
@@ -636,6 +636,20 @@
     ]
   },
   {
+    "component": "TabContainer",
+    "status": "alpha",
+    "source": "https://github.com/primer/view_components/tree/main/app/components/primer/alpha/tab_container.rb",
+    "lookbook": "https://primer.style/view-components/lookbook/inspect/primer/alpha/tab_container_preview/default/",
+    "parameters": [
+      {
+        "name": "system_arguments",
+        "type": "Hash",
+        "default": "N/A",
+        "description": "[System arguments](/system-arguments)"
+      }
+    ]
+  },
+  {
     "component": "TabNav",
     "status": "alpha",
     "source": "https://github.com/primer/view_components/tree/main/app/components/primer/alpha/tab_nav.rb",
@@ -2433,20 +2447,6 @@
         "default": "`false`",
         "description": "Whether to hide the border under the heading."
       },
-      {
-        "name": "system_arguments",
-        "type": "Hash",
-        "default": "N/A",
-        "description": "[System arguments](/system-arguments)"
-      }
-    ]
-  },
-  {
-    "component": "TabContainer",
-    "status": "alpha",
-    "source": "https://github.com/primer/view_components/tree/main/app/components/primer/tab_container_component.rb",
-    "lookbook": "https://primer.style/view-components/lookbook/inspect/primer/tab_container_preview/default/",
-    "parameters": [
       {
         "name": "system_arguments",
         "type": "Hash",
