+++ static/audited_at.json
@@ -27,6 +27,7 @@
   "Primer::Alpha::OcticonSymbols": "",
   "Primer::Alpha::SegmentedControl": "",
   "Primer::Alpha::SegmentedControl::Item": "",
+  "Primer::Alpha::TabContainer": "",
   "Primer::Alpha::TabNav": "",
   "Primer::Alpha::TabPanels": "",
   "Primer::Alpha::TextField": "",
