+++ static/statuses.json
@@ -27,6 +27,7 @@
   "Primer::Alpha::OcticonSymbols": "alpha",
   "Primer::Alpha::SegmentedControl": "alpha",
   "Primer::Alpha::SegmentedControl::Item": "alpha",
+  "Primer::Alpha::TabContainer": "alpha",
   "Primer::Alpha::TabNav": "alpha",
   "Primer::Alpha::TabPanels": "alpha",
   "Primer::Alpha::TextField": "alpha",
@@ -92,7 +93,7 @@
   "Primer::SpinnerComponent": "deprecated",
   "Primer::StateComponent": "deprecated",
   "Primer::SubheadComponent": "beta",
-  "Primer::TabContainerComponent": "alpha",
+  "Primer::TabContainerComponent": "deprecated",
   "Primer::TimeAgoComponent": "deprecated",
   "Primer::TimelineItemComponent": "beta",
   "Primer::TimelineItemComponent::BadgeComponent": "alpha",
