+++ test/components/alpha/tab_container_test.rb
@@ -2,17 +2,17 @@
 
 require "components/test_helper"
 
-class PrimerTabContainerComponentTest < Minitest::Test
+class PrimerAlphaTabContainerTest < Minitest::Test
   include Primer::ComponentTestHelpers
 
   def test_does_not_render_without_content
-    render_inline(Primer::TabContainerComponent.new)
+    render_inline(Primer::Alpha::TabContainer.new)
 
     refute_selector("tab-container")
   end
 
   def test_renders_content
-    render_inline(Primer::TabContainerComponent.new) { "content" }
+    render_inline(Primer::Alpha::TabContainer.new) { "content" }
 
     assert_selector("tab-container[data-view-component]", text: "content")
   end
