+++ test/components/component_test.rb
@@ -93,7 +93,7 @@ class PrimerComponentTest < Minitest::Test
     [Primer::Beta::Spinner, {}],
     [Primer::Beta::State, { title: "Open" }],
     [Primer::SubheadComponent, { heading: "Foo" }, proc { |component| component.with_heading { "Foo" } }],
-    [Primer::TabContainerComponent, {}, proc { "Foo" }],
+    [Primer::Alpha::TabContainer, {}, proc { "Foo" }],
     [Primer::Alpha::ToggleSwitch, {}],
     [Primer::Alpha::TextField, { name: :foo, label: "Foo" }],
     [Primer::Beta::Text, {}],
@@ -111,6 +111,7 @@ class PrimerComponentTest < Minitest::Test
 
   def test_registered_components
     ignored_components = [
+      "Primer::TabContainerComponent",
       "Primer::StateComponent",
       "Primer::OcticonSymbolsComponent",
       "Primer::SpinnerComponent",
